LPS-79878 Moving verifyContent to the upgrade to 6.2 since it was added by LPS-50555 at the beginning of that version. It is not needed for newer upgrades
